pyproject.toml
[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[project]
name = "neuro-news"
version = "0.1.0"
description = "AI-powered news terminal"
requires-python = ">=3.10"
dependencies = [
    "typer",
    "rich",
    "httpx",
    "feedparser",
    "platformdirs",
    "pydantic",
    "python-dateutil",
    "python-dotenv",
]

[project.scripts]
neuro-news = "neuro_news.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["neuro_news*"]
